+++ packages/carbon-graphs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Fixed reflow bug in timeline triggered by `showLegend=false`.
+  * Fixed retaining legend state in paired result graph during reflow.
   * Fixed isolation of different graph instances in same page.
 
 ## 2.17.0 - (April 27, 2021)